ansible/hg-ssh-server: use virtualenv as shebang in pash.py (bug 1329746); r=fubar
We didn't do this before as a protection against locking ourselves out
on CentOS 6. Now that we run CentOS 7 everywhere and pash is not
processing system SSH logins, we no longer need this protection and
can just use the shebang to the virtualenv in the file itself.

This helps reduce a window where pash is inconsistent and hg SSH
connections may fail due to using the wrong Python. But there's still
a window where the Python files are inconsistent amongst themselves.
A comment noting this has been added. (Fixing this requires moving
all the pash files into their own directory, which is a bit of scope
bloat for this bug.)

MozReview-Commit-ID: 7mxfmPfOMtK

--HG--
extra : amend_source : d5cdf63ba0f756279e08d24d91e299d451785424