robustcheckout: retry after "stream ended unexpectedly" (bug 1291926); r=glob
The most common network failure we've seen in automation is
"stream ended unexpectedly." In fact, this appears to be the only
network error we see with any frequency.

This commit introduces retry logic for network operations. The
default behavior is to attempt network pulls up to 3 times.

The code doesn't catch all network failures. But it catches the
big one: "stream ended unexpectedly," which is raised when
reading from changegroups, which constitute the bulk of bytes over
the wire for `hg pull` operations.

FWIW, I'm not sure of an exhaustive list of network related
exceptions Mercurial can emit. We'll likely have to follow up and
add more exception detection. Ideally, we'd implement retry logic
upstream. Perfect is the enemy of good.

MozReview-Commit-ID: 7fRBDNwus4w

--HG--
extra : rebase_source : 7816b49f97391e625722bf9818d99a023cb33b1f