overlay: command for overlaying a repo into a subdirectory of another (bug 1331697); r=glob
This commit introduces a Mercurial extension for applying changesets
from 1 Mercurial repository into the sub-directory of another.

The primary goal of this extension is to support ongoing vendoring
of "external" projects/repos into a monorepo in a way that preserves
history of the source repo while not convoluting the history of the
destination repo. The mechanisms used cater strongly to Mozilla's
repository management philosophy and the initial feature set is
somewhat limited because more advanced features are not yet
required by Mozilla.

The command works by cloning a "source" repository inside the
.hg directory of the destination. Requested changesets from that
source repository are then applied one at a time into the
destination repository. The files between the two changesets
are unioned.

The extension currently assumes that the destination directory
being written to is read-only outside of the overlay process and
will abort if there is file content mismatch. This property is
important because it prevents divergence between the source
repository and the directory in the destination. Divergence is
bad because, well, recovering from it is hard. When there is no
divergence, there is no potential for merge conflicts (which may
require human resolution) and there is no situation where the
diff of changeset X in the destination contains changes that
weren't there in the source repo. In the future, we may provide
a mechanism to "reconcile" the state of the destination directory
with the last overlayed changeset. Until then, read-only prevails.

The command keeps track of which changesets from the source have
been overlayed by storing metadata in the hidden "extras" dictionary
in changesets. Stored are the repository URL and revision the changeset
was overlayed from. This allows subsequent overlay operations to
find the last-overlayed changeset and quickly perform an
incremental overlay.

In theory, this metadata can be spoofed and an actor could
confuse the overlay mechanism. Mitigation for this is deferred
to server operators at this time. I'm not too concerned it will be
an issue at Mozilla in the short to medium term. I don't think there
are any security implications, since the metadata is only advisory
to aid subsequent overlay operations.

MozReview-Commit-ID: 7rKBb7xnU71

--HG--
extra : rebase_source : a6f0a25a295c8bf45e36ffec025fae9437085156