mozreview: change "previous/next commit" links to link to diffs from diff pages and reviews from review pages (bug 1330424). r=davidwalsh,smacleod
Changes the links to link to the next review on review pages and to the next diff on diff pages.
Previously it always went to review pages. This change makes it easier for reviewers to cycle
through the diffs without having to load an extra page.

MozReview-Commit-ID: J360HpxTIMm

--HG--
extra : rebase_source : bfa46bb943b280dbf033b746c2c1ff8ce5ff831e