robustcheckout: add test for server failure during clone (bug 1291926); r=glob
We're seeing a lot of "stream ended unexpectedly" failures when
cloning/pulling in automation. As far as I can tell, this is due
to dropped connections. Unreliable networks are unreliable.

This commit adds a test to reproduce the failure by implementing an
extension that wraps the server response serving function and has it
return early after a configured number of bytes have been sent.
This test will allow us to verify that automatric retry logic
(to be introduced in the next commit) works as advertised.

MozReview-Commit-ID: 96p8L7jDHeE

--HG--
extra : rebase_source : b6a46dc876072e231dba19f6316dde790ff601c6