ansible/hg-web: define liberal CSP policy for reftest analyzer (bug 1333929); r=fubar
Reftest analyzer loads images from "data:" URIs. Yuck. Until bug 1200501
is in place, we have to host the reftest analyzer on hg.mo.

This commit refactors the httpd code to define a separate, more
liberal CSP policy for the reftest analyzer from the rest of the site.

As part of this, the default CSP policy became more secure by dropping
some connect-src items!

There are multiple repositories with copies of the reftest analyzer.
However, an audit of the HTTP logs revealed 99+% of them hit the copy
in mozilla-central. So the URL check only looks for this instance.

The httpd config is a bit ugly. As the inline comments explain, this is
necessary because things like nested <If> and <Location> inside <If>
don't work.

Since the config code is now non-trivial and security is important,
I added a test for the CSP header.

MozReview-Commit-ID: 4MdsjRjrPam

--HG--
extra : rebase_source : ef6484e47bb138abd7c8ceef0ce2d4f9383de1c9