ansible/ius-repo: support installing on CentOS 7
This role was never updated to support CentOS 7 (likely because the
original reason for the repo was to obtain Python 2.7 on RHEL 6 and
CentOS 7). I have a new requirement for a package only provided in
IUS on CentOS 7. So let's support install IUS on CentOS 7.

MozReview-Commit-ID: Ib13UQvM3S9

--HG--
extra : rebase_source : 9a5ffdbb70e0e83c0040d352f92dffaddadaf98d