vcssync: support pushing converted Mercurial repo (bug 1337572); r=glob
As part of running a production service, we want to push the converted
Mercurial repo to another server as a means of backup. We already
support pushing the Git repo. So add functionality for pushing the
Mercurial repo.

MozReview-Commit-ID: BENunE1ZJWn

--HG--
extra : rebase_source : 23ac46bc4def0ed7a2fe39befba7d2935e504407