mozreview: Gracefully handle 500 errors from autoland server (bug 1338022); r=glob
If the autoland server throws an exception and generates an internal server
error, the AutolandTriggerResource would also throw an exception, as it
expected valid JSON in the response.  This in turn led to a cryptic
error being displayed to the user.

The entire response text is now returned as the error message if it is not
valid JSON.

MozReview-Commit-ID: KscP1RAnvgT

--HG--
extra : rebase_source : ccfc6e54b96da1cb2bb73cf6ca1520d3a3781b74