Bug 1342491 - Handle any kind of exception from git-cinnabar. r=glob
Bug 1298992 added an exception handler to avoid printing out a backtrace
on the common "Pushing merges is not supported yet" error.

The problem is that it checks e.args[0], but there is no guarantee that:
- 1) the exception args attribute is a non-empty list or tuple,
- 2) the exception has an args attribute at all,

The former is happening a lot lately, hiding the real error when people
are pushing to mozreview. The latter is more theoretical, in cases some
subclass of Exception doesn't forward its __init__ to Exception's.

This fixes both.

MozReview-Commit-ID: 842ay1oNZgl