testing: rsync from server instead of volume (bug 1349407); r=glob
For reasons that I can only attribute to "because Docker," building
Ansible Docker images is now extremely slow on my desktop. I tracked
this down to `rsync` being *extremely* slow. It was rsyncing between
a mounted Docker volume from another container and a local volume. I
put processes under `strace` and it revealed a number of lstat() and
lchown() calls. But they were extremely slow - milliseconds instead of
microseconds. I have no clue why this is or what changed to make this
slow. I'm using overlay2 if it matters.

Fortunately, there is a workaround which appears to fix the problem.

This commit switches the rsync from using a local volume to using the
rsync server running on the vct container. This makes rsync fast
again.

Various clobber.* files were touched to force the Docker images to be
rebuilt from scratch on next build. This is necessary to pick up
changes to the sync-and-build files from the base Docker image.

MozReview-Commit-ID: 2w1ZKbiFmrU

--HG--
extra : rebase_source : 2646dc1f6f718bba5b47e548a695aaf58e582970