generate-hg-s3-bundles: generate zstd bundles at level 21 (bug 1344790); r=glob
Level 22 was generating an out of memory error on 32-bit Windows.
Let's try level 21, which has a window size of half of level 22.
This will make the bundles a little larger. But, that's the price
we may have to pay for 32-bit compatbility.

MozReview-Commit-ID: AY3dkbuub9L

--HG--
extra : rebase_source : 3af257623d36fbb57ca5282ac92534fd8fcddeba