hgserver: create new repos as generaldelta by default (bug 1351848); r=glob
We previously had a config entry in the global hgrc files on hgssh and
hgweb that would result in new repos not using generaldelta. This was
initially done for performance reasons.

As I documented in the bug, I no longer think these performance issues
are a concern and it is safe to create new repos as generaldelta. So,
this commit removes the config option to default to the legacy repo
storage format.

Initially, a vcsreplicator test failed when I made the change. This
uncovered a bug in vcsreplicator which assumed new repos were created
as non-generaldelta by default. I've updated the vcsreplicator code
to explicitly specify a config option to control generaldelta usage
based as the named argument passed into a function. As part of this,
I switched to format.usegeneraldelta, which is semantically more
correct than format.generaldelta for repo creation.

MozReview-Commit-ID: HR6oKaYDs9T

--HG--
extra : rebase_source : 27036788ddb1b15f894bf22162e7f7fa41a24541