hgmo: delay import mozbuildinfo module; r=glob
Without this change, an upcoming commit that removes the bundleclone
extension from hgweb breaks the hgmo extension.

After much head scratching, I tracked this down to Mercurial's demand
importer and a bug in bundleclone.

The bundleclone extension uses the legacy, non-context-manager API
for disabling the demand importer when performing a module presence
test. Instead of restoring the previous state of the demand importer,
after this operation, it always calls "demandimport.enable()" thus
ensuring it is enabled.

When the bundleclone extension is removed, the hgmo extension is
loaded with the demand importer disabled. (We stopped enabling the
demand importer in WSGI processes in 287fd5985aa2.) This caused the
mozbuildinfo module import to fail because it attempts to import
mozbuild and mozpack modules which don't exist in the hgweb
virtualenv. This caused the hgmo module load to fail and hgweb
to effectively break.

The fix for this on hgmo's side is to delay import the module
containing unknown modules until they are actually needed. A
fix for bundleclone won't be pursued because the extension is
effectively deprecated and will be removed from use shortly.

MozReview-Commit-ID: E5LK1nfoA68

--HG--
extra : rebase_source : c46a7edc78f51b21cda5e4003705d8698640fa85
extra : histedit_source : 89944190cf1420c15ba6c6f1e5277dae7724987c