generate-hg-s3-bundles: use core command for producing stream clone bundles (bug 1352494); r=glob
The current behavior of generate-hg-s3-bundles is to call a command
in the bundleclone extension for producing a modern stream clone
bundle. If my recollection is correct, we implemented this functionality
in bundleclone to allow us to produce stream clone bundles before
deploying a new version of Mercurial that had built-in support for
this functionality.

Now that we have a modern version of Mercurial deployed, we don't need
to call the custom command implementation. So, this commit switches us
to the implementation in core Mercurial.

MozReview-Commit-ID: HI6CZYDwQ3h

--HG--
extra : rebase_source : 9586e50274481977509e2b108bf56abda3d4991b
extra : histedit_source : 44e36954f36f7b7c3702444951b4ae68d35bd3a7