ansible/hg-ssh: stop generating bzip2 bundles (bug 1354311); r=glob
The bzip2 bundles take a long time to generate and don't yield as good
compression as zstd. The bzip2 bundles exist to allow users to download
a smaller bundle. However, they aren't widely used because it requires
an explicit client-side config to prioritize them over gzip and zstd.

Since zstd is a better smaller bundle, let's remove the generation of
bzip2 bundles.

FWIW, I've already made this change on the server because of the
conversion of repos to generaldelta. The bundle generation code
uses ``bzip-v1`` as the bundle type. Using this against a generaldelta
repo would require several hours of CPU, which is why I disabled it.

MozReview-Commit-ID: BpG7JYpaCf6

--HG--
extra : rebase_source : f5138370d1f26c837ae697ecf7da5eccacbd9e5e