ansible/hg-web: emit CSP header from Mercurial (bug 1333615); r=glob
Mercurial 4.1 (which we just upgraded to) has built-in support for
emitting a Content-Security-Policy header. The big benefit over having
the parent HTTP server add the header is that Mercurial can generate
and use a nonce in inline <script> so the policy doesn't have to use
"unsafe-inline."

This commit moves the generation of the CSP header from httpd to hg.

As part of this, the CSP header was added to HTTP responses that
it previously wasn't. Despite me implementing the CSP feature
upstream, I forgot to special case protocol requests from emitting
the CSP policy like we did at Mozilla (to minimize bytes). Derp.
I may fix this in 4.2.

For now, let's pave over it by unsetting the header for identified
protocol requests. This required a clean-up block of sorts in the
vhost config. Apache httpd does have some wonky behavior when it
comes to order of execution of directives (especially when <If>
and <Location> are involved. But our tests running an actual
Apache httpd process using the same version used in production
show the desired behavior. So I'll take it.

MozReview-Commit-ID: EzTdIxING0W

--HG--
extra : rebase_source : 772709520e357eade13995ffe30bad0503e4dcbb