generate-hg-s3-bundles: renamed packed1-gd to packed1 (bug 1354311); r=glob
We no longer have generaldelta and non-generaldelta versions of these
bundles. So drop the "-gd" suffix since it isn't needed.

When we deploy this, we'll want to remove existing packed repo bundles
from the server to prevent any confusion arising from the rename.

As far as Mercurial is concerned, the filename doesn't matter: the only
thing that matters is the bundle specification in the clone bundles
manifest. And that doesn't change.

MozReview-Commit-ID: 1QJ1BzcMadz

--HG--
extra : rebase_source : 7cf88c1c15c1a0c47d3732c847f25fdce7cb29e2