robustcheckout: retry after socket errors (bug 1317594); r=glob
Previously, we didn't handle urllib.URLError and Mercurial's default
exception handler converted this to an error.Abort for display to the
user.

This commit explicitly catches urllib.URLError (aliased as
pycompat.urlerr.urlerror in Mercurial for compatibility with Python
3) and detects when the underlying error is a socket.error. It then
invokes the retry code for this scenario.

This should detect a whole class of network-related failures,
including connection refused, routing errors, and connection timeout.
Some of these might make sense to not retry on. But I'm inclined to
cast a wide net to ensure we aren't chasing the long tail of one-off
conditions that can contribute an intermittent network event.

This patch also doesn't handle non-socket errors in URLError. I'm
not as confident that we should retry on these. Let's see what happens
and we can evaluate things later.

MozReview-Commit-ID: JZ7mYBpnddD

--HG--
extra : rebase_source : 2940ce08fc7371b09aca1ac2dc830363a36a345b