ansible/hg-web: increase network timeout from 60s to 120s (bug 1291926); r=fubar
This should hopefully make many of the Mercurial client failures
reported in this bug go away. We had ~8000 of these "failed to proxy
response to client" errors in March. And the rate went up last week
when we converted various server repos to generaldelta. So we should
know relatively quickly if this change reduces the failure rate.

Currently, the load balancer is not enforcing an idle timeout on
connections. We should consider changing that. And once we do, we
can increase Timeout to effectively infinity, since as the in-line
comment explains, the thing it is measuring isn't terribly
important so it doesn't add much value.

MozReview-Commit-ID: AmsL7EZCnN6