robustcheckout: restore compatibility with Mercurial 3.7 (bug 1317594)
The recent changes to robustcheckout broke Mercurial 3.7. Our policy is
supposed to be only to support the last 4 releases of Mercurial. The
extension was previously updated to drop compatibility with 3.7 and to
explicitly require 3.8. However, parts of automation still use 3.7
(sadly).

The fix to robustcheckout is simply: reference urllib2.URLError directly
instead of through the pycompat module (which was introduced in 3.8).

The test suite now passes on 3.7.3 and all later versions.