global: support creating virtualenv for vcssync (bug 1357201); r=glob
Since the dawn of the version-control-tools repo we've shoehorned
everything into a single virtualenv. This worked for a while. But
the approach has been showing a number of cracks and has made it
difficult for some more isolated projects to exist in
version-control-tools. In one case (conduit), we created a new
repo partially because v-c-t's dev/test environment was... lacking.

This commit introduces the concept of multiple environments to
version-control-tools. Rather than a monolithic virtualenv, we can
create standalone virtualenvs.

The gateway to creating these environments is the `create-environment`
command. It accepts a name of the environment to create. Currently,
it only accepts "vcssync."

Creation and management of the new environments is implemented in
the vcttesting.environment module. The functionality is very primitive
and is basically some helper functions and inline "recipes." It gets
the job done.

The "vcssync" environment is defined by some pip requirements files,
which were added as part of this commit. There is a
"prod-requirements.txt" file defining the requirements to run vcssync
in production. We simply copy this file from Ansible (we'll remove
the redundancy in a later commit). "test-requirements.txt" defines
what is needed to run it locally. It simply includes
prod-requirements.txt and adds the coverage package (which is required
by run-tests).

A new documentation page has been created to show how all this works.

MozReview-Commit-ID: IIBSb8vUhAC

--HG--
rename : ansible/roles/vcs-sync/files/servo-requirements.txt => vcssync/prod-requirements.txt
extra : rebase_source : 71200a17081e978319215003ec84fbf2dbb9e729