testing: detect missing virtualenv differently (bug 1357201); r=glob
A consequence of adding new virtualenvs under venv/* is that venv/
may exist before ./create-{deploy,test}-environment is run. This was
throwing off virtualenv presence detection in the existing scripts.

We change the scripts to look for ${VENV}/lib instead. In addition,
a `mkdir` was changed to `mkdir -p` so the command won't fail if the
directory exists.

MozReview-Commit-ID: EedvaSPgsTp

--HG--
extra : rebase_source : 89cf919476891e9f81654d4edde65079cea13210