testing: take virtualenv into account when finding tests (bug 1357201); r=glob
Now that we can run tests with multiple virtualenvs and not all tests
can be executed with all virtualenvs, we need to declare which tests
are compatible with which virtualenvs. This commit implements that
logic.

Mercurial extension and hooks tests are now marked as only compatible
with the "global" virtualenv. All UNIT_TEST_DIRS are marked as
compatible with the "global" virtualenv except for vcssync/tests,
which requires the "vcssync" virtualenv.

This means if we have the "global" virtualenv activated, we won't be
able to run tests in vcssync/tests. That's fine: we want to decouple
vcssync from the global virtualenv. More importantly, it means that
if we have the "vcssync" virtualenv activated and run `run-tests`
that *only* the vcssync/tests will be executed. This is because no
other test file is discovered when this virtualenv is activated.

MozReview-Commit-ID: Grkl1kBN1Ir

--HG--
extra : rebase_source : 7dbbb53fc51e96fd514c9f4c7a1ed8646aed4c9b