vcssync: extract git commit rewriting to standalone function (bug 1357236); r=glob
linearize_git_repo() accepts a bunch of arguments and contains code
for rewriting Git commit object metadata (author field, summary line,
commit message, etc). This rewriting is not critical to the operation
of this function. It exists there for hysterical raisins.

As the inline TODO in that function states, we want to extract this
metadata rewriting to its own function so we have separation of
concerns and can perform similar rewriting from other consumers
(such as a Git pull request importer).

This commit extracts the Git commit metadata rewriting to its own
reusable function.

The new commit_metadata_rewriter() function accepts a repo and
behavioral arguments and returns a function that is called to
process a pair of old and new Git commit objects. This function
is then passed to linearize_git_repo() and called on each processed
commit.

This change is a bit larger than I would have liked. There was
significant churn around callers of linearize_git_repo(), including
from the CLI layer. Some of this was due to needing to construct a
dulwich.repo.Repo instance. Other changes were around handling of
argument passing. The new code relies much less on **kwargs and
I think that is a good thing because it is too easy for **kwargs
to hide subtle bugs like unused arguments. Unfortunately, it does
make the change a bit harder to review.

MozReview-Commit-ID: Azqmx5EkmBe

--HG--
extra : rebase_source : 5200e1d42532bdc771992789d58f513077e7ff26