mozreview: fix regression and remember try syntax from last run (bug 1357852). r=davidwalsh,glob
Bug 1228311 added a feature to remember the last try syntax that was
entered on a review request. Bug 1330205 modified the related code and
caused a regression by accidentaly removing part of the code that
supported the original change.

After this change the interface should remember the last try syntax
as describede in Bug 1228311. It doesn't seem possible to test this in
development though, since we are unable to mock landing end to end.

MozReview-Commit-ID: HwUXZ0vlgPc