vcssync: add utility function for obtaining a github3 client (bug 1358312); r=glob
We currently don't have test coverage of GitHub API requests. I'd like
to change that.

The first step is to create a function for obtaining a github3 client.
This will give us a central point from which we can inject special
behavior when running from tests.

While there is only one place where we obtain a github3 client today,
this will change in the near future. And the code will be sufficiently
complex to warrant a standalone function.

As part of this, I also snuck in a change to how the client is
constructed. ``github3.login()`` is equivalent to the new code,
at least for the usage we currently care about.

MozReview-Commit-ID: BZb26LeQPxk

--HG--
extra : rebase_source : 8baa2318e0f0ac46cc9632abb014d185a73503fd