vcssync: stop defining GitHub username (bug 1358312); r=glob
As part of refactoring code using github3 in a subsequent commit,
I realized that a username is not needed when using API tokens. The
code requires a username+password OR a token. Unless the password
is defined, the username argument is ignored.

So, stop defining and using a GitHub username value in the code and
in automation configs.

MozReview-Commit-ID: D3haUp0AyZx

--HG--
extra : rebase_source : f11fa7574b4eddd0279b72bb6b8a936ec9498eb6