mozreview: remove support for remotely stripping changesets (bug 1359945); r=mcote
This should be a rubber stamp review as far as the code goes. The "hard"
part of the review is deciding we want to do this.

The feature was implemented with the intent to automate stripping of
changesets. 2 years later and it isn't automated. Furthermore, this
feature isn't enabled in production because the remote_strip_users
config option isn't set. So the feature is dead code.

MozReview-Commit-ID: DD1pv0wYmxS