mozreview: always obtain a wlock (bug 1177742); r=glob
This is a somewhat hacky change.

Mercurial issues a warning when a wlock (working directory lock) is
obtained after a lock (store lock). This is to prevent deadlocks.

mercurial.exchange.push() always obtains a lock. However, it doesn't
always obtain a wlock. So when the reviewboard client extension goes
to rewrite changesets during push and obtains a wlock, it could be
obtaining a wlock after a lock and this could lead to deadlock.

We hack around this problem by setting a config option before calling
exchange.push() that will have the side-effect of causing Mercurial
to obtain a wlock. The config option in question enables the "pushback"
functionality, which enables a server to send rewritten changesets
back to the client. We don't use this feature on the server (it
requires a custom extension), so it should result in no major
change of behavior. But it will result in the client advertising
support for this feature to the server as part of the bundle2
capabilities string. Since the server ignores this, it should be no
problem.

The new code makes use of the ui.configoverride context manager.
This was added in Mercurial 4.1. This means our workaround only
works for 4.1+. I'm fine with that. Nobody has complained about
deadlock and the main purpose of this change is to allow us to
enable developer warnings in the test harness.

MozReview-Commit-ID: 5gceIvy6pIG

--HG--
extra : rebase_source : 180a1fb8b04c867ef099e8577e0afc5a6ccf6341