testing: add a requirement for vcsreplicator (bug 1360062); r=glob
We will be introducing a minimal dev environment for running Mercurial
tests. We want it to be able to execute tests in pylib/. However,
pylib/vcsreplicator currently contains a number of tests that
require vcsreplicator. And vcsreplicator has a number of dependencies
that we don't want to introduce to the generic and minimal hg dev
environment. So we want to skip them in this environment.

This commit introduces a new "vcsreplicator" requirement in hghave
to detect the presence of vcsreplicator and adds it to all
vcsreplicator tests. This will cause vcsreplicator tests to be
gracefully skipped in the minimal hg dev environment.

MozReview-Commit-ID: L8tQKkEQLhc

--HG--
extra : rebase_source : 4a5e32ad3c25195242f8c730906788a3db3faac4