testing: port Mercurial installation to Python (bug 1360062); r=glob
This essentially implements what `create-test-environment` does, just
in Python instead of shell. It allows `run-tests --all-versions` to work
with the hgdev virtualenv.

Since the functionality is redundant with what was in
`create-test-environment`, we add a shim to the Python module so the code
can be invoked standalone and we change `create-test-environment` to
call it. No DRY violation!

The global and hgdev environment share the Mercurial clone and
installation directory. I think this is better than giving each
environment its own install because the installed versions should
always be in sync and it cuts down on overhead.

MozReview-Commit-ID: 2AHwnABhaJe

--HG--
extra : rebase_source : dbe52599a24004b8973d8bd15e82da30bcc8b145