testing: generate coverage data files in own directory; r=glob
Coverage report files currently exist in the same directory as data
files. This can be a bit confusing and with a new version of coverage
even appears to emit a warning due to attempting to load a report file
as a data file.

This commit moves the coverage data files to their own directory so
they are isolated from non-data.

As part of this, I performed some light refactoring of run-tests.

MozReview-Commit-ID: EsFfQjbusbk

--HG--
extra : rebase_source : 83250c7eb1a59f8e7db0a016afdc22e5201be51b