testing: always set HGVERSION to resolved Mercurial version; r=glob
Previously, the "version" variable was the literal directory name. In
most cases, this should match exactly to the version reported by
`hg --version`. However, in the case of the "@" directory, the
version would be "@" which is nonsense.

We already have a helper function to resolve the Mercurial version
given an hg executable. So we plug that in to populate HGVERSION
with the actual version.

MozReview-Commit-ID: 59LtUX2sHnv

--HG--
extra : rebase_source : 979406801f2b68fd410ef6e1ca07f22cfe578967