hgext: mass upgrade extensions for Mercurial 4.2 compatibility (bug 1361182); r=glob
All extensions are now marked as compatibile with Mercurial 4.2.

As of this change, all non-Docker tests pass with both 4.1 and 4.2.

In some cases, the compatibility might be premature since Docker images
aren't yet running 4.2. I'm fine with that. We'll flush those out when
we upgrade the Docker images to 4.2.

MozReview-Commit-ID: EZggKVz1751

--HG--
extra : rebase_source : 3641d6fea3aa81e62fef58a6bcdef06567571a14