pushlog: drop compatibility with old Mercurial versions (bug 1361182); r=glob
The pushlog extension only needs to run on servers. I'm not sure why we
claim support for older versions. Probably a cargo cult (on my part).

This is causing a problem with tests because tests load the hgmo
extension (which has dropped compatibility <4.1) and this results in a
warning being printed on <4.1 Mercurials, causing tests to fail.

So let's mark the extensions with the same compatibility as the hgmo
extension so the tests don't run anywhere the hgmo extension isn't
also compatible.

MozReview-Commit-ID: JRp3bDv5cR1

--HG--
extra : rebase_source : b7b8d2a19574ad1e58e2254129c73ff49464e10f