testing: consistently name bmoweb environment variable (bug 1363509); r=glob
Environment variables are used by the test harness to pass the Docker
image IDs to tests. The environment variable name is typically
DOCKER_*_IMAGE where the wildcard is the name of the directory under
testing/docker. This naming was inconsistent for the bmoweb image,
so this commit fixes it.

MozReview-Commit-ID: HU1XtnJp1oo

--HG--
extra : rebase_source : 41079eb07a33cafeb3600f4bd0f17245e55ef51d