ansible: add role to send journald to CloudWatch (bug 1364231); r=dividehex
https://github.com/saymedia/journald-cloudwatch-logs is a Go program
for directly monitoring journald and writing events to Amazon CloudWatch
in near real time. It seems to be the most popular of the "send journald
to CloudWatch" tools I could find. Its compelling advantage over other
"send to CloudWatch" tools in this space is that it watches journald
directly: other tools poll files on disk. Because it is watching the
journal directly, it has access to the raw, unified stream of all
events. Not only does this provider greater assurances of handling
all the events, but it also doesn't result in data loss: the CloudWatch
events are JSON and contain everything from the raw journal message.
This facilitates easier analysis, including creating CloudWatch filters
to react to specific events.

This commit adds an Ansible role for configuring this tool. Things
are pretty straightforward: we have a user/group for running the
process via systemd. A minimal config file specifies which CloudWatch
log group to write to. The log name is the EC2 instance name by default.
I'm fine with this, at least to begin with. The documentation on GitHub
describes other configuration options.

To prove it works, we hook it up to the vcssync deployment.

MozReview-Commit-ID: KfMC7KdgFLu

--HG--
extra : rebase_source : bf32c1d98347b0fb4665d314a0bd5cc4e2dc3127
extra : amend_source : e35818dcce679fa64c868ea5acf51e4f2d92e8f7