docker: convert pulse image to CentOS 7
This is pretty straightforward.

CentOS doesn't have built-in packages for Erlang and RabbitMQ. So we
download RPMs published by the people who maintain Rabbit. We're not
running this in production, so I'm not too worried about SLA or
trust issues of the downloaded files.

And with this commit, all Docker images in the hg.mo environment are
now running from the same CentOS 7 base image! Yay consistency.

--HG--
extra : amend_source : 8726df4d50a27e9595736f245c23a95dde38dcc5