ansible: upgrade Mercurial in hgweb virtualenvs to 4.2 (bug 1360007); r=glob
We've loosened our Mercurial upgrade policy to allow taking the
initial X.Y release 2 weeks after release provided it looks stable.
There haven't been any major changes to the stable branch of the
Mercurial repo related to the 4.2 release. So I think it is stable
enough to go out on the hgweb servers.

Both hg-dev and hg-reviewboard were updated because we'll need to
update templates and the 4.2 templates aren't compatible with 4.1.
This is a bit unfortunate. But it should be relatively low risk.
There's a slight chance we'll break the reviewboard server extension.
But we have tests for that and tests are happy with this commit, so
everything should be fine.

The system package and virtualenvs for tools and replication are still
on 4.1. We'll upgrade those to 4.2 once we upgrade the rest of the
world to 4.2.

MozReview-Commit-ID: HcCEc7oqRCk

--HG--
extra : rebase_source : b027e48a3f7195901a28296f8fc7301de1ff3a27
extra : amend_source : 35762a104f09689d2b89dde6faa953571c979a46
extra : histedit_source : 5b13844c858b6f35ed9bd4435b5ae5d0cb78c315