testing: check for pip when determining if virtualenv exists (bug 1370694); r=glob
Previously, we'd look for a virtualenv directory when determining if
we needed to create a virtualenv. I'm not sure how, but periodically
I'll have an empty virtualenv directory and `create-environment` will
fail. So, change the check to look for the pip executable instead.

MozReview-Commit-ID: 6igydJ1Viwj

--HG--
extra : rebase_source : 3280e62f3c83ccca55fb0470293ed88073b52831