vcssync: support deleting refs in `git update-ref` helper (bug 1370694); r=glob
We'll need this in a subsequent commit. As part of this, I also
discovered that we weren't emitting a trailing null byte in the command
string. The Git man pages imply this is the proper format and sure
enough I did manage to get `git update-ref` to complain with it
missing and work with it present.

MozReview-Commit-ID: 3r46A4slEQv

--HG--
extra : rebase_source : 78a5a961a75815120744889ebc0dccda0e1e65f3