testing: have Mercurial use the virtualenv's Python
Currently, the Mercurial installed in the virtualenv uses the
virtualenv's Python and therefore has access to all packages installed
in the virtualenv.

The standalone Mercurial installs did not sure this property. Instead,
they likely used the system Python (or whatever Python was used to
create the virtualenv) in their shebang. This meant that Python
packages in the virtualenv were not exposed to these Mercurial
installs.

In this commit, we install the standalone Mercurials so they use
the virtualenv's Python.

This change unbreaks numerous failures in `run-tests --all-versions`
mode.

MozReview-Commit-ID: Cw0kVaPtfLH

--HG--
extra : histedit_source : b2b9203c7f1ca5ae5dd1d48990e075768d757118