testing: install Mercurial 4.1 explicitly in standalone mode
Currently, the virtualenv has a Mercurial package and we have various
standalone Mercurial installs. The virtualenv Mercurial is not
installed standalone. This is a bit cumbersome to maintain. The
main reason we don't install the virtualenv version in standalone
mode is redundancy: we don't want to run tests against the same
version multiple times during `run-tests --all-versions`.

I think a better solution is to just not run tests against the
virtualenv Mercurial when --all-versions is being used. The
next commit will implement that.

MozReview-Commit-ID: IaYRdIN73Ap

--HG--
extra : histedit_source : ef35b90b4843727db509723285fdc8807d626625