pushlog: add API to obtain push from node (bug 1360007); r=glob
Before, our API was to pass in a changectx. This was not as
flexible as it could be. Let's add a new API to accept a node and
rewrite pushfromchangeset() to call it.

Accepting a hex node would have been easier because everything pushlog
stores hex nodes. However, this is a bad design of pushlog: we should
operate with binary nodes wherever possible. I felt like drawing a
line in the sand.

This change does break some filelog tests in hgext/hgmo for an
obscure reason. The next commit will fix that.

MozReview-Commit-ID: 1WqPBvmWfXS

--HG--
extra : amend_source : 477a93b1bc274230d7a72320722a903ceef94c45
extra : histedit_source : 77cedf090223a568e9314cea7392d198a6e91201