vcsreplicator: use remote value for incoming bookmarks (bug 1268963); r=glob
Previously, `hg pull` relied on Mercurial's built-in behavior for
bookmark processing. The built-in behavior was to look at differences
between local and remote bookmarks and reconcile the differences.

Since the point of vcsreplicator is to run a mirror, the remote state
should be the final local state. So Mercurial's default behavior isn't
appropriate. This means we need to monkeypatch Mercurial.

We introduce a minimal consumer-only extension to be installed on
mirrors. The extension monkeypatches the bookmarks function that
is called when local bookmarks are to be updated from a remote.

The way bookmarks exchange works in Mercurial is that the full
set of bookmarks from the server is fetched and applied. In other
words, we never only receive bookmarks for just the changesets
that are being changed. This means it is sufficient to blow away
all local bookmarks and just let Mercurial's default behavior
kick in to re-set them. There may be some corner cases when
obsolescence is involved that this behavior breaks down. But
obsolescence is experimental. So I'd rather get bookmark replication
working for normal Mercurial and break obsolescence than spend a
lot of time writing tests that may not uncover a problem.

Writing this commit reminded me that I want to add an
`hg pull --mirror` feature to upstream Mercurial. I /think/ I have
buy-in on the idea. I just need to produce code.

MozReview-Commit-ID: Kkx9vcXW1fW

--HG--
extra : source : 577df9f966dc8fb72b6f6f02b53b792b5a8849bd
extra : amend_source : a5aa7772fd05d91ee59f65b8b52a2141f1e8df0b